requirements.txt
mysql-connector-python
youtube_dl